+++ Shared/DataModels/TemplateSheetFact.cs
@@ -23,7 +23,7 @@ public class TemplateSheetFact
     public string OpenRowSignature { get; set; }
     public string Unit { get; set; }
     public int Decimals { get; set; }
-    public decimal NumericValue { get; set; }
+    public double NumericValue { get; set; }
     public DateTime DateTimeValue { get; set; }
     public bool BooleanValue { get; set; }
     public string TextValue { get; set; }
@@ -128,7 +128,7 @@ public class TemplateSheetFact
                 try
                 {
                     var nfi = new CultureInfo("en-US", false).NumberFormat;
-                    NumericValue = Convert.ToDecimal(TextValue, nfi);
+                    NumericValue = Convert.ToDouble(TextValue, nfi);
                 }
                 catch (System.Exception)
                 {
